feat(timers): toggle by sort tag
